Support endpoint overrides (#32)
### Summary

Support endpoint overrides.

### Description

The Timestream for LiveAnalytics ODBC driver doesn't currently support endpoint overrides. An endpoint override can be set, but the only acceptable value is one of the service's regional public query endpoints. Attempting to use any other endpoint, such as a VPC endpoint, will fail, as the AWS SDK for C++ will call `DescribeEndpoints` using the overridden endpoint. This is due to limitations of the AWS SDK for C++.

As of AWS SDK for C++ version 1.11.663, endpoint overrides are supported for the `timestream-query` client. Now, when an endpoint override is set, endpoint discovery will be disabled and queries will be sent directly to the overridden endpoint.

In summary, to support overriding endpoints, this PR:
- Updates the AWS SDK for C++ dependency to 1.11.663.
- Updates the endpoint override integration test to actually test a practical endpoint override.
- Updates paths to AWS SDK for C++ libraries.
- Makes integers fixed width, regardless of platform, for the application data buffer.

- [X] Integration tests passed.
- [X] Unit tests passed.

### Related Issue

N/A.

### Additional Reviewers

N/A.